Add script to fill other languages
If there is no text for a language the key is displayed.
With the script you can copy a language (for example english) in a list of languages (for example korean).
By this way people running to the mod with a non supported language (korean in the example) will have mod text in the source language (in the example english) insteaf of non understanding localization keys.